Tweak several casts so that they work in both 32 and 64 bit

+++ compat/mingw.c
@@ -259,7 +259,7 @@ int pipe(int filedes[2])
 		CloseHandle(h[0]);
 		return -1;
 	}
-	fd = _open_osfhandle((int)h[0], O_NOINHERIT);
+	fd = _open_osfhandle((intptr_t)h[0], O_NOINHERIT);
 	if (fd < 0) {
 		close(filedes[0]);
 		close(filedes[1]);
@@ -269,7 +269,7 @@ int pipe(int filedes[2])
 	}
 	close(filedes[0]);
 	filedes[0] = fd;
-	fd = _open_osfhandle((int)h[1], O_NOINHERIT);
+	fd = _open_osfhandle((intptr_t)h[1], O_NOINHERIT);
 	if (fd < 0) {
 		close(filedes[0]);
 		close(filedes[1]);
